access PubChem to get geometries
